refactor: separate out common element styles
